Add remaining image elements, also add shake effect
I'm pretty proud of the shake effect, but it makes the scrolling of the background jitter weirdly, especially when you scroll across the train. I wonder if there's a way I could fix that or if I'm going to have to make the background simple enough that it won't be so noticeable.