add about page, clicking works
The background of the images was transparent so I thought I wouldn't be able to click on them, but that was not the case. I'm going to have to change the images of everything so they don't have all that transparent space I think, and that means having to painstakingly position them all properly :(

Also the about page is not formatted correctly yet, I need to center it, and I want to put like a zooming in transition between the pages so it looks like you're zooming up on illusioner. I get the feeling this will not be seamless though (page loading) so I might even end up trying to make the first page and the about information on the same page, meaning everything I have here might change later.