Start navi system in train
I was going to add the function of the arrows on the side to change the upcoming station and duplicate all of that to the other door too but I spent too long brainstorming station ideas in my sketchbook and now it's midnight whoops. In any case I have around 4 of them thought out so far, but a lot of my ideas are ambitious so none of it will be done for a while. My goal is to have at least an updateable blog page, some interaction with my V-Tor Graphica OCs, and some silly pages before the end of the summer.