add train stop (incomplete)
I made it so that the train animations (the shaking and moving background) stop once the name of the station is clicked on, and it registers the different stations through a single function, but there are a lot of bugs and issues. The animation is made to first increase the animation duration after pressing the name of the station so that it slows down before stopping, and that works for the first station, but it doesn't work for any of the others. Instead, it just pauses instantly and instantly changes stations too. I feel like it might have something to do with the event listener. Once I figure out how to fix that I'll add some blank webpages for actually going to the stations and duplicate the navigation code onto the right door as I planned. I'll work on the doors opening later since I think I'll make a gif of that or something and add it in there and I just want to get most of the code done for now.