Add blank station web pages, janky link to go to stations
I finally found time to work on this again, over a month since college started :'D I didn't fix any of the inconsistent train slowing or anything, as I want to hurry up and just make the website at least have the basics of all the features I want to include in it. Make it functional first, then make it good later, as I say. I'll go in an pretty up everything and reorganize later, since I need to add and crop all the drawing elements of the website anyways (you cannot click through the transparent parts of images, I found out.) I ended up putting an a element the size of the door on top of the door drawing, which gains the link to the correct station webpage after the train stops.