Add minor adjustments to text, complete navigation system
I adjusted the text in the about section to be a little easier to understand, and made navigating between the different dialogue options controlled by 1 function instead of 3 to make it more efficient. Similarly, I improved the efficiency of the navigation system in the part where you change the target station by having just one a element where clicking the arrows changes the innerHTML to the correct station name. I had to make these changes to make it easier to copy the navigation functionality to the other set of doors on the right.

I also got rid of the slowing to a stop thing that I was trying to do with the train when the target station was chosen. I couldn't figure out how to make it work consistently and I realized it would also be a little annoying to deal with when navigating the website anyways. So at this point I have the functionality of the homepage complete, though it will always be open to changes later.